+++ src/react/src/components/InitialClaimFlow/ClaimForm/styles.js
@@ -1,0 +1,85 @@
+import COLOURS from '../../../util/COLOURS';
+import { HEADER_HEIGHT } from '../../../util/constants';
+
+const claimFormStyles = theme =>
+    Object.freeze({
+        container: Object.freeze({
+            backgroundColor: '#F9F7F7',
+            padding: '48px 0 120px 0',
+        }),
+        innerContainer: Object.freeze({
+            maxWidth: '896px',
+            margin: '0 auto',
+        }),
+        paper: Object.freeze({
+            padding: '40px 40px',
+            backgroundColor: COLOURS.WHITE,
+            boxShadow: 'none',
+            borderRadius: 0,
+            marginTop: '32px',
+        }),
+        title: Object.freeze({
+            marginBottom: '48px',
+            fontSize: '56px',
+            fontWeight: 900,
+            lineHeight: '60px',
+            textAlign: 'center',
+            color: COLOURS.JET_BLACK,
+        }),
+        description: Object.freeze({
+            marginBottom: '24px',
+            fontSize: '18px',
+            fontWeight: 600,
+            textAlign: 'center',
+        }),
+        navigationButtons: Object.freeze({
+            justifyContent: 'space-between',
+            padding: '48px 0',
+        }),
+        buttonBack: Object.freeze({
+            width: '200px',
+            height: '49px',
+            borderRadius: 0,
+            textTransform: 'none',
+            fontSize: '18px',
+            fontWeight: theme.typography.fontWeightExtraBold,
+            border: '1px solid #0D1128',
+        }),
+        buttonPrimary: Object.freeze({
+            width: 'fit-content',
+            height: '49px',
+            borderRadius: 0,
+            textTransform: 'none',
+            backgroundColor: theme.palette.action.main,
+            color: theme.palette.common.black,
+            fontSize: '18px',
+            fontWeight: theme.typography.fontWeightExtraBold,
+            boxShadow: 'none',
+            '&:hover': {
+                backgroundColor: theme.palette.action.dark,
+            },
+            '&:disabled': {
+                backgroundColor: COLOURS.GREY,
+                color: COLOURS.DARK_GREY,
+                cursor: 'not-allowed',
+            },
+        }),
+        sectionTitle: Object.freeze({
+            fontSize: '1.25rem',
+            fontWeight: 600,
+            marginBottom: theme.spacing.unit * 2,
+        }),
+        sectionDescription: Object.freeze({
+            fontSize: '0.875rem',
+            color: COLOURS.DARK_GREY,
+            marginBottom: theme.spacing.unit * 3,
+        }),
+        loadingContainer: Object.freeze({
+            display: 'flex',
+            justifyContent: 'center',
+            alignItems: 'center',
+            height: `calc(100vh - ${HEADER_HEIGHT}px)`,
+        }),
+    });
+
+export default claimFormStyles;